Create "executeGame" and "gameBoard" functions.
"executeGame" is the main function. Acts as the game controller,
reacting to the user's tic-tac-toe grid interactions to execute the
different game functions in the corresponding order.

"gameBoard" uses the tic-tac-toe grid (retieved in "executeGame") to
create an updated board "currentBoard" object reflecting the status of
the game and updating after each turn (user selection).

A test tic-tac-toe grid was created in html to help building the logic.